moved loader to `phichain-game`
