+++ phichain-game/Cargo.toml
@@ -37,3 +37,5 @@ bevy_prototype_lyon = "0.11"
 num = "0.4.3"
 rand = "0.8.5"
 image = { version = "0.24", features = ["jpeg", "png"] }
+anyhow = "1.0.86"
+serde_json = "1.0.117"
